ADMIN

+++ src/database/migrations/versions/2023_04_06_0958-2fd5dbd9e541_add_book_table.py
@@ -21,7 +21,7 @@ def upgrade() -> None:
     op.create_table('book',
     sa.Column('name', sa.String(), nullable=False),
     sa.Column('description', sa.String(), nullable=False),
-    sa.Column('price', sa.Numeric(), nullable=False),
+    sa.Column('price', sa.Float(), nullable=False),
     sa.Column('stock', sa.Integer(), nullable=False),
     sa.Column('category_id', sa.Integer(), nullable=False),
     sa.Column('store_id', sa.Integer(), nullable=False),
